fix(ci): resolve full CI red across toolchain, clippy, and infra gaps
Root causes fixed, all pre-existing and unrelated to any single PR:

- libs/api: c_variadic feature-gate kept portable across toolchain versions
  (deleting it broke older nightlies still requiring the flag)
- virtio_drivers: lifetime-elision warning
- deny.toml: cargo-deny schema drift — unmaintained "all"/"workspace" scope
  string replaced with "deny" (matches original enforcement intent, just
  fixes the outdated syntax); GPL-2.0-only exception renamed to the bare
  SPDX id the current schema expects
- dco.yml: broken action reference (dco-org/check-dco does not exist and
  was never caught, so DCO sign-off has never actually been enforced on
  this repo); replaced with an inline non-interactive check (no third-party
  Action) that verifies every commit in the PR range carries a
- ci.yml Build matrix + geiger-gate: littlefs2-sys (via redoxfs) needs a
  matching cross C compiler per target that was never installed/wired
- ci.yml: timeout-minutes guards on apt-get steps after two jobs hung 88-100
  minutes on a transient package-mirror stall

Clippy (-D warnings) cleanup across libs/api, hal-arm, hal-x86, and the
kernel crate on all three targets (aarch64, x86_64, riscv64): missing
# Safety docs on posix/HAL unsafe fns, arch-specific dead code caused by
iommu_riscv.rs/iommu_x86.rs/other arch modules being unconditionally
compiled on every target (now cfg-gated to their real target, matching
the existing per-call cfg pattern in iommu.rs), plus mechanical lints
(redundant casts, manual div_ceil/is_multiple_of, needless index loops,
missing Default impls, etc.). Escalated fixes requiring judgment (not
mechanical) were verified individually:

- result_unit_err (task.rs, 20 fns): suppressed with a documented reason
  rather than redesigning error types across call sites
- declare_interior_mutable_const (hart_local.rs, irq_wait.rs, etc.):
  verified these are the correct pattern for per-slot array-repeat atomics,
  not the aliasing bug the lint warns about; suppressed with reason
- not_unsafe_ptr_arg_deref (hypervisor::registry::run_vcpu): marked
  unsafe fn, call site in task/syscall.rs updated to match (already
  validated by validate_user_buf beforehand)

Also fixed two vendored third_party/redoxfs warnings (unused import,
unused variable) needed to keep the vendored tree compiling under -D
warnings on every target.

Verified: full 54-test boot-suite green after regen, aarch64/x86_64/riscv64
`-p vicell-kernel` builds clean under -D warnings, aarch64+x86_64 clippy
clean under -D warnings.

+++ libs/api/src/services/posix/strings.rs
@@ -5,6 +5,11 @@
 
 use core::ffi::{c_char, c_int, c_void};
 
+/// Copy `n` bytes from `src` to `dest`.
+///
+/// # Safety
+/// `dest` and `src` must each be valid for `n` bytes of reads/writes respectively, and the
+/// two regions must not overlap (use `memmove` if they might).
 #[no_mangle]
 pub unsafe extern "C" fn memcpy(dest: *mut c_void, src: *const c_void, n: usize) -> *mut c_void {
     let d = dest as *mut u8;
@@ -15,6 +20,11 @@ pub unsafe extern "C" fn memcpy(dest: *mut c_void, src: *const c_void, n: usize)
     dest
 }
 
+/// Copy `n` bytes from `src` to `dest`, correctly handling overlapping regions.
+///
+/// # Safety
+/// `dest` and `src` must each be valid for `n` bytes of reads/writes respectively. Overlap
+/// between the two regions is permitted.
 #[no_mangle]
 pub unsafe extern "C" fn memmove(dest: *mut c_void, src: *const c_void, n: usize) -> *mut c_void {
     let d = dest as *mut u8;
@@ -33,6 +43,10 @@ pub unsafe extern "C" fn memmove(dest: *mut c_void, src: *const c_void, n: usize
     dest
 }
 
+/// Fill the first `n` bytes of `s` with the low byte of `c`.
+///
+/// # Safety
+/// `s` must be valid for `n` bytes of writes.
 #[no_mangle]
 pub unsafe extern "C" fn memset(s: *mut c_void, c: c_int, n: usize) -> *mut c_void {
     let d = s as *mut u8;
@@ -43,6 +57,10 @@ pub unsafe extern "C" fn memset(s: *mut c_void, c: c_int, n: usize) -> *mut c_vo
     s
 }
 
+/// Compare the first `n` bytes of `s1` and `s2`.
+///
+/// # Safety
+/// `s1` and `s2` must each be valid for `n` bytes of reads.
 #[no_mangle]
 pub unsafe extern "C" fn memcmp(s1: *const c_void, s2: *const c_void, n: usize) -> c_int {
     let s1 = core::slice::from_raw_parts(s1 as *const u8, n);
@@ -54,6 +72,11 @@ pub unsafe extern "C" fn memcmp(s1: *const c_void, s2: *const c_void, n: usize) 
     0
 }
 
+/// Return the length of the NUL-terminated string `s`, excluding the terminator.
+///
+/// # Safety
+/// `s` must point to a valid, NUL-terminated string readable up to and including the
+/// terminator.
 #[no_mangle]
 pub unsafe extern "C" fn strlen(s: *const c_char) -> usize {
     let mut len = 0;
@@ -61,6 +84,11 @@ pub unsafe extern "C" fn strlen(s: *const c_char) -> usize {
     len
 }
 
+/// Copy the NUL-terminated string `src` into `dest`, including the terminator.
+///
+/// # Safety
+/// `src` must be a valid NUL-terminated string. `dest` must be valid for writes of at least
+/// `strlen(src) + 1` bytes. The two buffers must not overlap.
 #[no_mangle]
 pub unsafe extern "C" fn strcpy(dest: *mut c_char, src: *const c_char) -> *mut c_char {
     let mut i = 0;
@@ -73,6 +101,13 @@ pub unsafe extern "C" fn strcpy(dest: *mut c_char, src: *const c_char) -> *mut c
     dest
 }
 
+/// Copy at most `n` bytes from the NUL-terminated string `src` into `dest`, zero-padding any
+/// remaining bytes when `src` is shorter than `n`.
+///
+/// # Safety
+/// `src` must be a valid NUL-terminated string readable up to its terminator or `n` bytes,
+/// whichever comes first. `dest` must be valid for writes of `n` bytes. The two buffers must
+/// not overlap.
 #[no_mangle]
 pub unsafe extern "C" fn strncpy(dest: *mut c_char, src: *const c_char, n: usize) -> *mut c_char {
     let mut i = 0;
@@ -88,29 +123,44 @@ pub unsafe extern "C" fn strncpy(dest: *mut c_char, src: *const c_char, n: usize
     dest
 }
 
+/// Lexicographically compare NUL-terminated strings `s1` and `s2`.
+///
+/// # Safety
+/// `s1` and `s2` must each be valid NUL-terminated strings readable up to their terminator.
 #[no_mangle]
 pub unsafe extern "C" fn strcmp(s1: *const c_char, s2: *const c_char) -> c_int {
     let mut i = 0;
     loop {
-        let c1 = *s1.add(i) as u8;
-        let c2 = *s2.add(i) as u8;
+        let c1 = *s1.add(i);
+        let c2 = *s2.add(i);
         if c1 != c2 { return c1 as c_int - c2 as c_int; }
         if c1 == 0 { return 0; }
         i += 1;
     }
 }
 
+/// Compare at most `n` bytes of the NUL-terminated strings `s1` and `s2`.
+///
+/// # Safety
+/// `s1` and `s2` must each be readable up to their terminator or `n` bytes, whichever comes
+/// first.
 #[no_mangle]
 pub unsafe extern "C" fn strncmp(s1: *const c_char, s2: *const c_char, n: usize) -> c_int {
     for i in 0..n {
-        let c1 = *s1.add(i) as u8;
-        let c2 = *s2.add(i) as u8;
+        let c1 = *s1.add(i);
+        let c2 = *s2.add(i);
         if c1 != c2 { return c1 as c_int - c2 as c_int; }
         if c1 == 0 { return 0; }
     }
     0
 }
 
+/// Append the NUL-terminated string `src` to the end of the NUL-terminated string `dest`.
+///
+/// # Safety
+/// `dest` must be a valid NUL-terminated string with enough trailing capacity to hold
+/// `strlen(dest) + strlen(src) + 1` bytes. `src` must be a valid NUL-terminated string. The
+/// two buffers must not overlap.
 #[no_mangle]
 pub unsafe extern "C" fn strcat(dest: *mut c_char, src: *const c_char) -> *mut c_char {
     let len = strlen(dest);
@@ -118,30 +168,45 @@ pub unsafe extern "C" fn strcat(dest: *mut c_char, src: *const c_char) -> *mut c
     dest
 }
 
+/// Find the first occurrence of byte `c` in the NUL-terminated string `s`, returning null if
+/// not found (the search includes the terminator when `c == 0`).
+///
+/// # Safety
+/// `s` must be a valid NUL-terminated string readable up to its terminator.
 #[no_mangle]
 pub unsafe extern "C" fn strchr(s: *const c_char, c: c_int) -> *mut c_char {
     let v = c as u8;
     let mut i = 0;
     loop {
-        let b = *s.add(i) as u8;
+        let b = *s.add(i);
         if b == v { return s.add(i) as *mut c_char; }
         if b == 0 { return core::ptr::null_mut(); }
         i += 1;
     }
 }
 
+/// Find the last occurrence of byte `c` in the NUL-terminated string `s`, returning null if
+/// not found (the search includes the terminator when `c == 0`).
+///
+/// # Safety
+/// `s` must be a valid NUL-terminated string readable up to its terminator.
 #[no_mangle]
 pub unsafe extern "C" fn strrchr(s: *const c_char, c: c_int) -> *mut c_char {
     let v = c as u8;
     let len = strlen(s);
     let mut i = len;
     loop {
-        if *s.add(i) as u8 == v { return s.add(i) as *mut c_char; }
+        if *s.add(i) == v { return s.add(i) as *mut c_char; }
         if i == 0 { return core::ptr::null_mut(); }
         i -= 1;
     }
 }
 
+/// Find the first occurrence of byte `c` within the first `n` bytes of `s`, returning null if
+/// not found.
+///
+/// # Safety
+/// `s` must be valid for `n` bytes of reads.
 #[no_mangle]
 pub unsafe extern "C" fn memchr(s: *const c_void, c: c_int, n: usize) -> *mut c_void {
     let v = c as u8;
